requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
av==16.0.1
babel==2.17.0
beautifulsoup4==4.14.2
blake3==1.0.5
bleach==6.2.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
compressed-tensors==0.10.2
contourpy==1.3.2
courlan==1.3.2
crcmod==1.7
cryptography==44.0.3
cycler==0.12.1
dateparser==1.2.2
ddgs==9.9.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
duckduckgo_search==8.1.1
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fake-useragent==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.2
fastrlock==0.8.3
filelock==3.19.1
fonttools==4.60.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
google_search_results==2.4.2
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
ImageIO==2.37.2
interegular==0.3.3
ipykernel==7.0.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jusText==3.0.2
kiwisolver==1.4.9
lark==1.2.2
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
lxml==6.0.1
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
mistune==3.1.4
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
narwhals==2.14.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
notebook==7.4.7
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.0
openai==1.99.5
openai-harmony==0.0.4
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
outlines_core==0.2.10
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
primp==0.15.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
qwen-vl-utils==0.0.14
ray==2.49.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.37.1
serpapi==0.1.5
setproctitle==1.3.7
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
stack-data==0.6.3
starlette==0.47.3
streamlit==1.52.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.11.0
tinycss2==1.4.0
tld==0.13.1
tokenizers==0.22.0
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
typer==0.17.4
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
wadllib==1.3.6
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
yarl==1.20.1
zipp==3.23.0

